fix(board): fixed crash when drag/drop from grid square to another grid square
fixed bug where dragging from one square on the playing grid to another square would crash the app.
Feature is still buggy as tiles do not render while dragging from the grid, but they will move and
state is updated properly. Moving a tile from one square to another square with a tile already on it
will swap the positions of both tiles in the state and on the board. Placing a tile from the player
tiles into the square with a tile already on it will overwrite the existing tile, and logic for
dragging a tile from the board back into the player pile is not yet implemented.